requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
drf-writable-nested==0.6.3
idna==3.3
oauthlib==3.2.0
phonenumberslite==8.12.45
Pillow==9.0.1
psycopg2-binary==2.9.3
pyshiki==1.1.7
pyshikiapi==1.0.1
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
